Added code formatting and linting tools, updated the code to fix issues.
