+++ package.json
@@ -3,12 +3,18 @@
   "version": "0.0.1",
   "private": true,
   "devDependencies": {
+    "eslint": "^4.18.2",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-config-react": "^1.1.7",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.7.0",
+    "prettier": "^1.11.1",
     "react-scripts": "1.1.1"
   },
   "dependencies": {
     "bootstrap": "^3.3.7",
-    "eslint": "^4.18.2",
-    "prettier": "^1.11.1",
     "react": "^16.2.0",
     "react-bootstrap": "^0.32.1",
     "react-dom": "^16.2.0",
